requirements.txt
pyserial>=3.5
pyautogui>=0.9.54
pydirectinput>=1.0.4
pystray>=0.19.5
